feat: Add juju3 support for func tests.
Add make target, tox environment to support juju3 along with juju2.

Some changes were done to run the same set of func tests for both
libjuju versions without having to check which version we're on:

* Replace `run_wait()` method with `run_command_on_unit()` in
conftest.py's `SyncHelper` class. This changes the functionality from
running a command on a unit with libjuju's `unit.run` (which could
potentially return different dictionaries for juju2 and juju3), to a
similar `juju exec --unit` command run using ops_test.juju.

* `juju exec` return code is different for failure between juju2 and
juju3. So the assert check for failure is changed to "> 0" instead of
checking strict equality with a non-zero integer.