fix: Remove redfish credential validation in install event handler (#192)
The exporter.check_active method being called in the restart_exporter method (from the update status event handler) will fail in case the exporter installation has not been done. This installation process takes place as part of the install event handler.

The error happens because the check_installed function in service.py (called as a decorator for the various exporter service helper methods) would return False when the exporter isn't installed. This would cause all the Exporter class' methods in service.py (restart, check_active, check_health etc.) to return False. The issue occurs when the on_update_status() --> restart_exporter() raises an exception when exporter.check_active consistently returns False.

This commit removes the redfish credential validation step during the install event since it's likely to not have the right configuration during the installation. Moreover, this can leave the charm in an inconsistent state where the exporter is not properly installed, leading to an error during the execution of other lifecycle events.