tests: Add hw dependent tests for redfish cred validation (#204)
Adds `test_redfish_credential_validation` which sets the redfish
credentials and checks if the charm is funtioning as intended.

`test_build_and_deploy` is broken up into smaller tests so that they
are easier to manage. Also modifies logic in conftest.py so that all
these tests are executed whether they're hw dependent or independent.

`test_redfish_metrics` has also been modified since we are setting the
redfish credentials as part of the earlier test.

Finally, a retry mechanism has been added to `test_metrics_available` as
it was found while testing that it takes some time for the endpoint to
be ready with the metrics.