Add a custom metrics CSV for the dcgm snap (#325)
Provides a custom metrics CSV file for the dcgm snap. The file can be provided through the dcgm-exporter-metrics-file configuration. The default metrics from the upstream are not comprehensive, so I added other ones that can be helpful to monitor in production.

Configures the dcgm-exporter to use the custom metrics file.