Wait for microk8s to be ready (#434)
* Wait for microk8s to be ready

I suspect that microk8s charm can be active/idle, but the k8s
cluster might not be ready. Running the job I was constant hitting
problems to create the clusterroles for creating the juju_k8s_cloud.
Running again terraform the job was able to create the k8s cloud

Adding the `microk8s status --wait-ready` is a more reliable check
and fixed the issue.

* - apply suggestions
- used terraform fmt

* - re-added deleted line

* - add and remove a clusterrole to check that kube-apiserver is ready

---------

Co-authored-by: Samuel Allan <samuel.allan@canonical.com>